requirements-kaggle-p100.txt
# Kaggle P100 (Linux) recommended versions.
# Install torch from the official CUDA wheel index.
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

transformers==4.39.3
accelerate==0.27.2
numpy==1.26.4
pillow==10.2.0
pyyaml==6.0.1
tqdm==4.66.2
matplotlib==3.8.3
